Feedback on IQ score

Provides feedback for the adaptive test framed in terms of an 'IQ' score. IQ scores are rescaling of IRT scores to a scale with mean 100 and standard deviation 15. The feedback also provides an interpration of the IQ score as a percentile with respect to the general population.

cat.feedback.iq(test_label, text = "feedback_iq",
  dict = psychTestRCAT::ptrcat_dict, next_button = NULL,
  digits_iq = 0L, digits_percentile = 0L)

Arguments

test_label
(Character scalar) The label for the test, e.g. "MDT". This will be used to label the IQ score, e.g. "Your MDT-IQ was...".

text
(Character scalar) Text to display to the participant. This text will be treated as an internationalisation key for the dictionary contained in the dict argument; if the key is not present in the dictionary, the text will be displayed as is.

dict
Internationalisation dictionary (see i18n_dict).

next_button
(NULL or a character scalar or an object of class "shiny.tag") If NULL, no next button is shown (typically because the test has completed). Otherwise, a button to progress to the next page is created, displaying the content of this argument.

digits_iq
(Numeric scalar) Number of digits to which the IQ score should be rounded.

digits_percentile
(Numeric scalar) Number of digits to which the percentile score should be rounded.

feedback_iq
(Character scalar) Text to display to the participant. This text will be treated as an internationalisation key for the dictionary contained in the dict argument; if the key is not present in the dictionary, the text will be displayed as is.

Value

A timeline segment suitable for inclusion in a psychTestR timeline directly after adapt_test.