Compile trial-by-trial results

Compiles a data frame of trial-by-trial results for each participant. This function is intended to be called once data have been collected for several participants; it allows the researcher to inspect test results on a trial-by-trial level.

compile_trial_by_trial_results(in_dir = "output/results", label,
  combine = TRUE)

Arguments

in_dir
Results directory to process.

label
Label that the test's results were saved under (e.g. "MDT").

combine
Whether to combine results into one big data frame, or instead to return a list of data frames, one for each participant.

Value

A data frame, or list of data frames, of trial-by-trial results.